requirements.txt
Django==3.2.9
python-dotenv==0.19.2
python-telegram-bot==13.8.1

qrcode~=7.3.1
Pillow==8.4.0
phonenumbers~=8.12.37
dateutils==0.6.12

whitenoise~=5.3.0
dj-database-url~=0.5.0
gunicorn~=20.1.0
psycopg2==2.8.6